requirements.txt
streamlit
google-generativeai
python-dotenv
python-docx
pdfplumber
langchain
genai
langchain-google-genai